Remove RecordProcessorAnnouncementHelper
